Refactor: added the types to Types folder and made checklogged in user resuable
