+++ bond-frontend/app/(social)/social/Types/UserDataTypes.ts
@@ -1,0 +1,10 @@
+export type UserData = {
+  createdAt: string;
+  email: string;
+  firstname: string;
+  lastname: string;
+  updatedAt: string;
+  username: string;
+  _v: number;
+  _id: string;
+};
